Refactor: MVVM으로 리팩토링
